// helpers/smartContractsWizard/interfaces/smartContractInfo.ts
import { SmartContractStandard } from "../utils/smartContractStandards.js";
export interface SmartContractInfo {
	name: string;
	symbol: string;
	standard: SmartContractStandard;
	isMintable: boolean;
	isBurnable: boolean;
	isPausable: boolean;
	isVotes: boolean;
	isOwnable: boolean;
	isRoles: boolean;
}